+++ R/pccc.R
@@ -23,7 +23,7 @@
   # intersection, and the union. Build the matrices separately so tech-only
   # signals can be handled after checking whether a non-tech condition was
   # flagged for the patient.
-  any_transplant <- unique(mdcr_subset(x = cmrb, i = cmrb[["transplant_flag"]] == 1L, cols = c(id.vars)))
+  any_transplant <- mdcr_unique(mdcr_subset(x = cmrb, i = cmrb[["transplant_flag"]] == 1L, cols = c(id.vars)))
 
   # Build ccc indicators based on only non-tech_dep codes
   # _or_ a tech_dep transplant.  There are codes which are both tech_dep and
@@ -144,7 +144,7 @@
   if (subconditions) {
     rtn <- list(conditions = rtn, subconditions = list())
     for (cnd in conditions) {
-      scnd <- unique(mdcr_subset(cmrb, i = cmrb[["condition"]] == cnd, cols = c(id.vars, "subcondition")))
+      scnd <- mdcr_unique(mdcr_subset(cmrb, i = cmrb[["condition"]] == cnd, cols = c(id.vars, "subcondition")))
 
       # subset the uiddf to the set of those with the condition
       uiddf <- mdcr_subset(rtn[["conditions"]],
@@ -177,11 +177,11 @@
 #' @noRd
 #' @keywords internal
 .pccc_v2 <- function(id.vars, iddf, cmrb, subconditions) {
-  any_tech_dep   <- unique(mdcr_subset(cmrb, i = cmrb[["tech_dep_flag"]] == 1L,   cols = c(id.vars)))
-  any_transplant <- unique(mdcr_subset(cmrb, i = cmrb[["transplant_flag"]] == 1L, cols = c(id.vars)))
+  any_tech_dep   <- mdcr_unique(mdcr_subset(cmrb, i = cmrb[["tech_dep_flag"]] == 1L,   cols = c(id.vars)))
+  any_transplant <- mdcr_unique(mdcr_subset(cmrb, i = cmrb[["transplant_flag"]] == 1L, cols = c(id.vars)))
 
   CMRB <- mdcr_select(cmrb, cols = c(id.vars, "condition"))
-  CMRB <- unique(CMRB)
+  CMRB <- mdcr_unique(CMRB)
 
   conditions <- sort(unique(..mdcr_internal_pccc_conditions..[["condition"]]))
 
@@ -214,7 +214,7 @@
   if (subconditions) {
     rtn <- list(conditions = rtn, subconditions = list())
     for (cnd in conditions) {
-      scnd <- unique(mdcr_subset(cmrb, i = cmrb[["condition"]] == cnd, cols = c(id.vars, "subcondition")))
+      scnd <- mdcr_unique(mdcr_subset(cmrb, i = cmrb[["condition"]] == cnd, cols = c(id.vars, "subcondition")))
 
       # subset the uiddf to the set of those with the condition
       uiddf <- mdcr_subset(rtn[["conditions"]],